Fixed azure tests
The Azure storage tests were using a “transfer-encoding: chunked”
header in the mock response as a way to avoid having to actually send
back the body contents (getFile just waits for the response, not the
body).  Those tests also added a content-length header to the response.

There was an http parser bug fix in Node that has been merged into the
most recent releases of maintained versions of Node (including 0.10.x)
which treats the simultaneous presence of a content-length header and a
“transfer-encoding: chunked” header as an error (see
https://github.com/nodejs/node/commit/375f35514c51fa603dae9fac86bb61c050
3cb678).

The fix for the azure tests was to remove the “transfer-encoding:
chunked” headers from the mock response, and to make the mock responses
to the GET return the actual file contents (with content-length header).

Verified with 0.10.43, 4.3.2, and 5.8.0.