[fix] Patched RegEx DoS vuln & Remote Memory Exposure vuln This addresses two major security issues:
https://nodesecurity.io/advisories/309
https://nodesecurity.io/advisories/535

First, a note: before adding this commit, master branch had 18 failing tests. I've ensured that after this commit the same 18 tests are failing, and not any different ones.

This patch addresses vulnerabilities with `mime` and `request` modules being used.

`mime` v1.x was recently discovered to be vulnerable to a RegEx Denial-of-Serice attack. This has two vectors within `pkgcloud` as it includes `mime` directly, and the old version of `request` being pinned is also using a vulnerable version of `mime`.

`request` also had its own vulnerability, a Remote Memory Exposure vulnerability.

Unfortunately just updating the versions on these resulted in more failing tests. The test dependency, `hock`, does not work as expected with `request` v2.83.0 due to some differences with how it expects JSON to be exposed for pattern matching. I've adjusted the tests with a workaround (heavy escaping), however I will also submit a pull request to the `hock` repository with a fix for this.